update csproj file
