+++ src/Bard.gRPC/Bard.gRPC.csproj
@@ -1,0 +1,25 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <Nullable>enable</Nullable>
+        <PackageId>Bard.gRPC</PackageId>
+        <Authors>Stuart Clark</Authors>
+        <Company>Stuart Clark</Company>
+        <PackageTags>BDD Testing Fluent gRPC API micro services</PackageTags>
+        <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+        <PackageVersion />
+        <Description>A .NET Functional gRPC API Testing Framework</Description>
+        <PackageProjectUrl>https://docs.bard.net.nz</PackageProjectUrl>
+        <RepositoryUrl>https://github.com/sjclark76/Bard</RepositoryUrl>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Bard\Bard.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Grpc.Core.Api" Version="2.27.0" />
+        <PackageReference Include="Grpc.Net.Client" Version="2.27.0" />
+    </ItemGroup>
+</Project>
