+++ src/Bard.gRPC/GrpcScenarioContext.cs
