+++ src/Bard.gRPC/IGrpcWhen.cs
