+++ src/Bard.sln
@@ -8,6 +8,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bard", "Bard\Bard.csproj", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bard.gRPCService", "Bard.gRPCService\Bard.gRPCService.csproj", "{CCADC441-A881-4C31-BB30-F4F10644A723}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bard.gRPC", "Bard.gRPC\Bard.gRPC.csproj", "{7D3DF7CB-44EC-4256-9086-2F83D9AB0B74}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -30,5 +32,9 @@ Global
 		{CCADC441-A881-4C31-BB30-F4F10644A723}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CCADC441-A881-4C31-BB30-F4F10644A723}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CCADC441-A881-4C31-BB30-F4F10644A723}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7D3DF7CB-44EC-4256-9086-2F83D9AB0B74}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7D3DF7CB-44EC-4256-9086-2F83D9AB0B74}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7D3DF7CB-44EC-4256-9086-2F83D9AB0B74}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7D3DF7CB-44EC-4256-9086-2F83D9AB0B74}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
