+++ src/Bard/Bard.csproj
@@ -17,12 +17,14 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Grpc.Core.Api" Version="2.27.0" />
-        <PackageReference Include="Grpc.Net.Client" Version="2.27.0" />
         <PackageReference Include="Microsoft.AspNetCore.WebUtilities" Version="2.2.0" />
         <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.6" />
         <PackageReference Include="MinVer" Version="2.3.0" />
         <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="gRPC" />
     </ItemGroup>
 
 </Project>
